+++ lib/menu_screen.dart
@@ -21,8 +21,8 @@ class SideDrawer<T> extends StatefulWidget {
     this.footerView,
     this.selectedItemId,
     this.slide = false,
-    double percentage,
-    double degree,
+    double? percentage,
+    double? degree,
     this.onMenuItemSelected,
     this.child,
     this.color = Colors.white,
@@ -40,7 +40,7 @@ class SideDrawer<T> extends StatefulWidget {
     this.itemBuilder,
     this.elevation = 16,
     this.cornerRadius,
-    Key key,
+    Key? key,
   })  : assert((child != null && menu == null && itemBuilder == null) ||
             (child == null && menu != null)),
         this.percentage = percentage ?? 0.8,
@@ -63,10 +63,10 @@ class SideDrawer<T> extends StatefulWidget {
   final double elevation;
 
   /// Card's corner radius
-  final double cornerRadius;
+  final double? cornerRadius;
 
   /// Degree of rotation : 15->45 degree
-  final double degree;
+  final double? degree;
 
   /// Drawer's width in Pixel,
   /// Default : 300px
@@ -78,17 +78,17 @@ class SideDrawer<T> extends StatefulWidget {
 
   /// Transition's [Curve],
   /// Default : [Curves.easeOut]
-  final Curve curve;
+  final Curve? curve;
 
   /// Transition's [Duration],
   /// Defalut: 250ms
-  final Duration duration;
+  final Duration? duration;
 
   /// [Menu] for drawer
-  final Menu<T> menu;
+  final Menu<T>? menu;
 
   /// Current selected ID
-  final T selectedItemId;
+  final T? selectedItemId;
 
   /// Flag for animation on menu item
   final bool animation;
@@ -97,35 +97,35 @@ class SideDrawer<T> extends StatefulWidget {
   final bool slide;
 
   /// listen to menu selected
-  final Function(T) onMenuItemSelected;
+  final Function(T)? onMenuItemSelected;
 
   /// [Widget] for header on drawer
-  final Widget headerView;
+  final Widget? headerView;
 
   /// [Widget] for footer on drawer
-  final Widget footerView;
+  final Widget? footerView;
 
   /// Custom builder for menu item
-  final MenuItemBuilder<T> itemBuilder;
+  final MenuItemBuilder<T>? itemBuilder;
   // final SideDrawerItemBuilder itemBuilder;
 
   /// Widget for side drawer
-  final Widget child;
+  final Widget? child;
 
   /// Background for drawer
-  final DecorationImage background;
+  final DecorationImage? background;
 
   /// Background graident for drawer
-  final LinearGradient gradient;
+  final LinearGradient? gradient;
 
   /// Background [Color] for drawer
   final Color color;
 
   /// [Color] for selected menu item
-  final Color selectorColor;
+  final Color? selectorColor;
 
   /// Menu item [TextStyle]
-  final TextStyle textStyle;
+  final TextStyle? textStyle;
 
   /// Menu [Alignment] in drawer
   final Alignment alignment;
@@ -146,23 +146,23 @@ class SideDrawer<T> extends StatefulWidget {
   final Curve slideInCurve;
 
   double maxSlideAmount(context) =>
-      drawerWidth ?? MediaQuery.of(context).size.width * percentage;
+      drawerWidth; // ?? MediaQuery.of(context).size.width * percentage;
 
   @override
   _SideDrawerState createState() => _SideDrawerState();
 }
 
 class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
-  double selectorYTop;
-  double selectorYBottom;
+  double? selectorYTop;
+  double? selectorYBottom;
 
-  Color selectorColor;
-  TextStyle textStyle;
+  Color? selectorColor;
+  TextStyle? textStyle;
 
   double get maxSlideAmount => widget.maxSlideAmount(context);
 
   setSelectedRenderBox(RenderBox newRenderBox, bool useState) async {
-    final renderBox = context.findRenderObject() as RenderBox;
+    final renderBox = context.findRenderObject() as RenderBox?;
 
     final newYTop =
         newRenderBox.localToGlobal(Offset(0.0, 0.0), ancestor: renderBox).dy;
@@ -188,23 +188,23 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
     final List<Widget> listItems = [];
 
     if (widget.child != null) {
-      listItems.add(widget.child);
+      listItems.add(widget.child ?? SizedBox());
     } else {
       final animationIntervalDuration = 0.5;
       final perListItemDelay =
           menuController.state != MenuState.closing ? 0.15 : 0.0;
       final millis = menuController.state != MenuState.closing
-          ? 150 * widget.menu.items.length
+          ? 150 * widget.menu!.items.length
           : 600;
 
-      final maxDuration = (widget.menu.items.length - 1) * perListItemDelay +
+      final maxDuration = (widget.menu!.items.length - 1) * perListItemDelay +
           animationIntervalDuration;
 
-      for (var i = 0; i < widget.menu.items.length; ++i) {
+      for (var i = 0; i < widget.menu!.items.length; ++i) {
         final animationIntervalStart = i * perListItemDelay;
         final animationIntervalEnd =
             animationIntervalStart + animationIntervalDuration;
-        MenuItem item = widget.menu.items[i];
+        MenuItem item = widget.menu!.items[i];
         listItems.add(buildListItem(menuController, item,
             animationIntervalStart, animationIntervalEnd, millis, maxDuration));
       }
@@ -246,7 +246,7 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
     final isSelected = item.id == widget.selectedItemId;
 
     Function onTap = () {
-      widget.onMenuItemSelected(item.id);
+      widget.onMenuItemSelected!(item.id);
       menuController.close();
     };
     Widget listItem = widget.itemBuilder == null
@@ -256,22 +256,27 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
             title: item.title,
             isSelected: isSelected,
             selectorColor: selectorColor,
-            textStyle: textStyle,
+            textStyle: item.textStyle ?? textStyle,
             menuView: widget,
             width: maxSlideAmount,
-            icon: item.icon == null ? null : item.icon.call(),
-            onTap: onTap,
+// <<<<<<< HEAD
+//             icon: item.icon == null ? null : item.icon.call(),
+//             onTap: onTap,
+// =======
+            icon: item.icon == null ? item.prefix : item.icon!.call(),
+            suffix: item.suffix,
+            onTap: onTap as dynamic Function()?,
             drawBorder: !widget.animation,
           )
         : InkWell(
             child: Container(
               alignment: Alignment.centerLeft,
               child: Container(
-                child: widget.itemBuilder(context, item, isSelected),
+                child: widget.itemBuilder!(context, item, isSelected),
                 width: maxSlideAmount,
               ),
             ),
-            onTap: onTap,
+            onTap: onTap as void Function()?,
           );
 
     if (widget.animation)
@@ -282,7 +287,7 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
         curve: Interval(animationIntervalStart / maxDuration,
             animationIntervalEnd / maxDuration,
             curve: Curves.easeOut),
-        menuListItem: listItem,
+        menuListItem: listItem as _MenuListItem?,
       );
     else {
       return listItem;
@@ -329,15 +334,36 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
 
   @override
   Widget build(BuildContext context) {
-    selectorColor = widget?.selectorColor ?? Theme.of(context).indicatorColor;
-    textStyle = widget?.textStyle ??
-        Theme.of(context).textTheme.subtitle1.copyWith(
+    selectorColor = widget.selectorColor ?? Theme.of(context).indicatorColor;
+    textStyle = widget.textStyle ??
+        Theme.of(context).textTheme.subtitle1?.copyWith(
             color: widget.color.computeLuminance() < 0.5
                 ? Colors.white
                 : Colors.black);
     return DrawerScaffoldMenuController(
         direction: widget.direction,
         builder: (BuildContext context, MenuController menuController) {
+          var shouldRenderSelector = true;
+          var actualSelectorYTop = selectorYTop;
+          var actualSelectorYBottom = selectorYBottom;
+          var selectorOpacity = 1.0;
+
+          if (menuController.state == MenuState.closed ||
+              menuController.state == MenuState.closing ||
+              selectorYTop == null) {
+            final RenderBox? menuScreenRenderBox =
+                context.findRenderObject() as RenderBox?;
+
+            if (menuScreenRenderBox != null) {
+              final menuScreenHeight = menuScreenRenderBox.size.height;
+              actualSelectorYTop = menuScreenHeight - 50.0;
+              actualSelectorYBottom = menuScreenHeight;
+              selectorOpacity = 0.0;
+            } else {
+              shouldRenderSelector = false;
+            }
+          }
+
           return Container(
             // padding: widget.direction == Direction.right
             //     ? const EdgeInsets.only(left: 24)
@@ -361,10 +387,8 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
         });
   }
 
-  static _SideDrawerState of(BuildContext context, {bool nullOk = true}) {
-    assert(nullOk != null);
-    assert(context != null);
-    final _SideDrawerState result =
+  static _SideDrawerState? of(BuildContext context, {bool nullOk = true}) {
+    final _SideDrawerState? result =
         context.findAncestorStateOfType<_SideDrawerState>();
     if (nullOk || result != null) return result;
     throw FlutterError.fromParts(<DiagnosticsNode>[
@@ -376,12 +400,12 @@ class _SideDrawerState extends State<SideDrawer> with TickerProviderStateMixin {
 }
 
 class ItemSelector extends ImplicitlyAnimatedWidget {
-  final double top;
-  final double bottom;
-  final double left;
-  final double opacity;
+  final double? top;
+  final double? bottom;
+  final double? left;
+  final double? opacity;
 
-  final Color selectorColor;
+  final Color? selectorColor;
 
   ItemSelector({
     this.left,
@@ -396,39 +420,39 @@ class ItemSelector extends ImplicitlyAnimatedWidget {
 }
 
 class _ItemSelectorState extends AnimatedWidgetBaseState<ItemSelector> {
-  Tween<double> _topY;
-  Tween<double> _bottomY;
-  Tween<double> _opacity;
+  Tween<double?>? _topY;
+  Tween<double?>? _bottomY;
+  Tween<double?>? _opacity;
 
   @override
-  void forEachTween(TweenVisitor visitor) {
+  void forEachTween(visitor) {
     _topY = visitor(
       _topY,
-      widget.top,
+      widget.top!,
       (dynamic value) => Tween<double>(begin: value),
-    );
+    ) as Tween<double?>?;
     _bottomY = visitor(
       _bottomY,
-      widget.bottom,
+      widget.bottom!,
       (dynamic value) => Tween<double>(begin: value),
-    );
+    ) as Tween<double?>?;
     _opacity = visitor(
       _opacity,
-      widget.opacity,
+      widget.opacity!,
       (dynamic value) => Tween<double>(begin: value),
-    );
+    ) as Tween<double?>?;
   }
 
   @override
   Widget build(BuildContext context) {
     return Positioned(
-      top: _topY.evaluate(animation),
+      top: _topY!.evaluate(animation),
       left: widget.left,
       child: Opacity(
-        opacity: _opacity.evaluate(animation),
+        opacity: _opacity!.evaluate(animation)!,
         child: Container(
           width: 5.0,
-          height: _bottomY.evaluate(animation) - _topY.evaluate(animation),
+          height: _bottomY!.evaluate(animation)! - _topY!.evaluate(animation)!,
           color: widget.selectorColor,
         ),
       ),
@@ -437,18 +461,23 @@ class _ItemSelectorState extends AnimatedWidgetBaseState<ItemSelector> {
 }
 
 class AnimatedMenuListItem extends ImplicitlyAnimatedWidget {
-  final Widget menuListItem;
-  final MenuState menuState;
-  final bool isSelected;
+// <<<<<<< HEAD
+//   final Widget menuListItem;
+//   final MenuState menuState;
+//   final bool isSelected;
+// =======
+  final _MenuListItem? menuListItem;
+  final MenuState? menuState;
+  final bool? isSelected;
   final Duration duration;
 
   AnimatedMenuListItem({
     this.menuListItem,
     this.menuState,
     this.isSelected,
-    this.duration,
-    Curve curve,
-    Key key,
+    required this.duration,
+    required Curve curve,
+    Key? key,
   }) : super(key: key, duration: duration, curve: curve);
 
   @override
@@ -460,20 +489,20 @@ class _AnimatedMenuListItemState
   final double closedSlidePosition = 200.0;
   final double openSlidePosition = 0.0;
 
-  _SideDrawerState get _sideDrawerState => _SideDrawerState.of(context);
+  _SideDrawerState? get _sideDrawerState => _SideDrawerState.of(context);
 
-  Tween<double> _translation;
-  Tween<double> _opacity;
+  Tween<double?>? _translation;
+  Tween<double?>? _opacity;
 
   updateSelectedRenderBox(bool useState) {
-    final renderBox = context.findRenderObject() as RenderBox;
-    if (renderBox != null && widget.isSelected) {
-      _sideDrawerState?.setSelectedRenderBox?.call(renderBox, useState);
+    final renderBox = context.findRenderObject() as RenderBox?;
+    if (renderBox != null && widget.isSelected!) {
+      _sideDrawerState?.setSelectedRenderBox.call(renderBox, useState);
     }
   }
 
   @override
-  void forEachTween(TweenVisitor visitor) {
+  void forEachTween(visitor) {
     var slide, opacity;
 
     switch (widget.menuState) {
@@ -487,19 +516,22 @@ class _AnimatedMenuListItemState
         slide = openSlidePosition;
         opacity = 1.0;
         break;
+
+      default:
+        break;
     }
 
     _translation = visitor(
       _translation,
       slide,
       (dynamic value) => Tween<double>(begin: value),
-    );
+    ) as Tween<double?>?;
 
     _opacity = visitor(
       _opacity,
       opacity,
       (dynamic value) => Tween<double>(begin: value),
-    );
+    ) as Tween<double?>?;
   }
 
   @override
@@ -507,11 +539,11 @@ class _AnimatedMenuListItemState
     updateSelectedRenderBox(false);
 
     return Opacity(
-      opacity: _opacity.evaluate(animation),
+      opacity: _opacity!.evaluate(animation)!,
       child: Transform(
         transform: Matrix4.translationValues(
           0.0,
-          _translation.evaluate(animation),
+          _translation!.evaluate(animation)!,
           0.0,
         ),
         child: widget.menuListItem,
@@ -522,42 +554,44 @@ class _AnimatedMenuListItemState
 
 class _MenuListItem extends StatelessWidget {
   final String title;
-  final bool isSelected;
-  final bool drawBorder;
-  final Function() onTap;
-  final Color selectorColor;
-  final TextStyle textStyle;
-  final SideDrawer menuView;
-  final Widget icon;
+  final bool? isSelected;
+  final bool? drawBorder;
+  final Function()? onTap;
+  final Color? selectorColor;
+  final TextStyle? textStyle;
+  final SideDrawer? menuView;
+  final Widget? icon;
+  final Widget? suffix;
   final Direction direction;
-  final double width;
-  final EdgeInsets padding;
+  final double? width;
+  final EdgeInsets? padding;
 
   _MenuListItem({
-    this.title,
+    required this.title,
     this.isSelected,
     this.onTap,
     this.menuView,
-    @required this.textStyle,
-    @required this.selectorColor,
+    required this.textStyle,
+    required this.selectorColor,
     this.icon,
     this.drawBorder,
     this.direction = Direction.right,
     this.width,
     this.padding,
+    this.suffix,
   });
 
   @override
   Widget build(BuildContext context) {
-    TextStyle _textStyle =
-        textStyle.copyWith(color: isSelected ? selectorColor : textStyle.color);
+    TextStyle _textStyle = textStyle!
+        .copyWith(color: isSelected! ? selectorColor : textStyle!.color);
 
     List<Widget> children = [];
     if (icon != null)
       children.add(Padding(
         padding: EdgeInsets.only(right: 12),
         child: IconTheme(
-            data: IconThemeData(color: _textStyle.color), child: icon),
+            data: IconThemeData(color: _textStyle.color), child: icon!),
       ));
     children.add(
       Expanded(
@@ -570,24 +604,31 @@ class _MenuListItem extends StatelessWidget {
         flex: 1,
       ),
     );
+    if (suffix != null)
+      children.add(Padding(
+        padding: EdgeInsets.only(right: 12),
+        child: IconTheme(
+            data: IconThemeData(color: _textStyle.color), child: suffix!),
+      ));
     return InkWell(
       splashColor: const Color(0x44000000),
-      onTap: isSelected ? null : onTap,
+      onTap: isSelected! ? null : onTap,
       child: Container(
         width: width,
         alignment: Alignment.centerRight,
-        // padding: padding,
-        decoration: drawBorder
+        decoration: drawBorder!
             ? ShapeDecoration(
                 shape: Border(
                   left: BorderSide(
-                      color: isSelected ? selectorColor : Colors.transparent,
+                      color: isSelected == true
+                          ? selectorColor!
+                          : Colors.transparent,
                       width: 5.0),
                 ),
               )
             : null,
         child: Padding(
-          padding: menuView.padding,
+          padding: menuView!.padding,
           child: Row(
             mainAxisAlignment: MainAxisAlignment.start,
             children: children,
@@ -602,32 +643,59 @@ class Menu<T> {
   final List<MenuItem<T>> items;
 
   const Menu({
-    this.items,
+    required this.items,
   });
 }
 
 typedef IconBuilder = Widget Function();
 
 class MenuItem<T> {
-  final T id;
+  final T? id;
   final String title;
-  final IconBuilder icon;
+  final IconBuilder? icon;
+
+  /// set icon from [MenuItem], if the icon is not null, the prefix must be null
+  // final IconData? icon;
+
+  /// set prefix widget from [MenuItem], if the prefix is not null, the icon must be null
+  final Widget? prefix;
+
+  /// set prefix widget from [MenuItem]
+  final Widget? suffix;
+
+  /// set independent text style for title
+  final TextStyle? textStyle;
+
+  /// append data with [MenuItem], then can be use on itemBuilder
+  final dynamic data;
 
   MenuItem({
     this.id,
-    this.title,
+    required this.title,
     this.icon,
-  });
+    this.prefix,
+    this.suffix,
+    this.textStyle,
+    this.data,
+  }) : assert(prefix == null || icon == null);
 
   MenuItem<T> copyWith({
-    T id,
-    String title,
-    IconBuilder icon,
+    T? id,
+    String? title,
+    IconBuilder? icon,
+    Widget? prefix,
+    Widget? suffix,
+    TextStyle? textStyle,
+    dynamic? data,
   }) {
     return MenuItem<T>(
       id: id ?? this.id,
       title: title ?? this.title,
       icon: icon,
+      prefix: prefix,
+      suffix: suffix,
+      textStyle: textStyle ?? this.textStyle,
+      data: data ?? this.data,
     );
   }
 }